feat(#216): Book room modal
